put aux methods into a separate class
