+++ samza-test/src/test/java/org/apache/samza/test/processor/TestZkStreamProcessor.java
@@ -19,37 +19,10 @@
 
 package org.apache.samza.test.processor;
 
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
-import java.util.Properties;
 import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
-import kafka.utils.TestUtils;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.samza.config.ApplicationConfig;
-import org.apache.samza.config.Config;
-import org.apache.samza.config.JobCoordinatorConfig;
-import org.apache.samza.config.MapConfig;
-import org.apache.samza.config.ZkConfig;
 import org.apache.samza.processor.StreamProcessor;
-import org.apache.samza.processor.StreamProcessorLifecycleListener;
-import org.apache.samza.system.IncomingMessageEnvelope;
-import org.apache.samza.system.OutgoingMessageEnvelope;
-import org.apache.samza.system.SystemStream;
-import org.apache.samza.task.InitableTask;
-import org.apache.samza.task.MessageCollector;
-import org.apache.samza.task.StreamTask;
-import org.apache.samza.task.TaskContext;
-import org.apache.samza.task.TaskCoordinator;
-import org.apache.samza.test.StandaloneIntegrationTestHarness;
-import org.apache.samza.test.StandaloneTestUtils;
 import org.apache.samza.zk.TestZkUtils;
 import org.junit.Assert;
 import org.junit.Before;
@@ -59,7 +32,7 @@ import org.junit.Test;
  * Happy path tests.
  * Start 1, 2, 5 processors and make sure they all consume all the events.
  */
-public class TestZkStreamProcessor extends StandaloneIntegrationTestHarness {
+public class TestZkStreamProcessor extends TestZkStreamProcessorBase {
 
   private final static int ATTEMPTS_NUMBER = 5; // to avoid long sleeps, we rather use multiple attempts with shorter sleeps
   
@@ -67,7 +40,6 @@ public class TestZkStreamProcessor extends StandaloneIntegrationTestHarness {
   public void setupTest() {
 
   }
-
 
   @Test
   public void testSingleStreamProcessor() {
@@ -98,7 +70,7 @@ public class TestZkStreamProcessor extends StandaloneIntegrationTestHarness {
     createTopics(inputTopic, outputTopic);
 
     // create a latch of the size == number of messages
-    TestStreamTask.endLatch = new CountDownLatch(messageCount);
+    TestZkStreamProcessorBase.TestStreamTask.endLatch = new CountDownLatch(messageCount);
 
     // initialize the the processors
     // we need startLatch to know when the processor has been completely initialized
@@ -115,7 +87,7 @@ public class TestZkStreamProcessor extends StandaloneIntegrationTestHarness {
     // run the processors in a separate threads
     Thread[] threads = new Thread[processorIds.length];
     for (int i = 0; i < processorIds.length; i++) {
-      threads[i] = runInThread(streamProcessors[i], TestStreamTask.endLatch);
+      threads[i] = runInThread(streamProcessors[i], TestZkStreamProcessorBase.TestStreamTask.endLatch);
       threads[i].start();
       // wait until the processor reports that it has started
       try {
@@ -127,7 +99,7 @@ public class TestZkStreamProcessor extends StandaloneIntegrationTestHarness {
 
     // wait until all the events are consumed
     try {
-      TestStreamTask.endLatch.await(10, TimeUnit.SECONDS);
+      TestZkStreamProcessorBase.TestStreamTask.endLatch.await(10, TimeUnit.SECONDS);
     } catch (InterruptedException e) {
       Assert.fail("endLatch.await failed with an interruption:" + e.getLocalizedMessage());
     }
@@ -346,7 +318,7 @@ public class TestZkStreamProcessor extends StandaloneIntegrationTestHarness {
     // wait until p2 consumes all the message by itself;
     attempts = ATTEMPTS_NUMBER;
     while (attempts > 0) {
-      long leftEventsCount = TestStreamTask.endLatch.getCount();
+      long leftEventsCount = TestZkStreamProcessorBase.TestStreamTask.endLatch.getCount();
       System.out.println("2current count = " + leftEventsCount);
       if (leftEventsCount == 0) { // should read all of them
         System.out.println("2read all. current count = " + leftEventsCount);
@@ -372,191 +344,5 @@ public class TestZkStreamProcessor extends StandaloneIntegrationTestHarness {
     // but the expected values are the same 0-79 - we should get each value one time.
     // Meanwhile the number of events we gonna get is 40 + 80
     verifyNumMessages(outputTopic, 2 * messageCount, totalEventsToGenerate);
-  }
-
-
-  //////////////////////////////////////////////////////////////////////////////////////////////////
-  // auxiliary methods
-  private StreamProcessor createStreamProcessor(final String pId, Map<String, String> map,
-      final CountDownLatch startLatchCountDown, final CountDownLatch stopLatchCountDown) {
-    map.put(ApplicationConfig.PROCESSOR_ID, pId);
-
-    StreamProcessor processor = new StreamProcessor(new MapConfig(map), new HashMap<>(), TestStreamTask::new,
-        new StreamProcessorLifecycleListener() {
-
-          @Override
-          public void onStart() {
-            if (startLatchCountDown != null) {
-              startLatchCountDown.countDown();
-            }
-          }
-
-          @Override
-          public void onShutdown() {
-            if (stopLatchCountDown != null) {
-              stopLatchCountDown.countDown();
-            }
-            System.out.println("ON STOP. PID = " + pId + " in thread " + Thread.currentThread());
-
-          }
-
-          @Override
-          public void onFailure(Throwable t) {
-
-          }
-        });
-
-    return processor;
-  }
-
-  private void createTopics(String inputTopic, String outputTopic) {
-    TestUtils.createTopic(zkUtils(), inputTopic, 5, 1, servers(), new Properties());
-    TestUtils.createTopic(zkUtils(), outputTopic, 5, 1, servers(), new Properties());
-  }
-
-  private Map<String, String> createConfigs(String testSystem, String inputTopic, String outputTopic,
-      int messageCount) {
-    Map<String, String> configs = new HashMap<>();
-    configs.putAll(StandaloneTestUtils
-            .getStandaloneConfigs("test-job", "org.apache.samza.test.processor.TestZkStreamProcessor.TestStreamTask"));
-    configs.putAll(StandaloneTestUtils
-        .getKafkaSystemConfigs(testSystem, bootstrapServers(), zkConnect(), null, StandaloneTestUtils.SerdeAlias.STRING,
-            true));
-    configs.put("task.inputs", String.format("%s.%s", testSystem, inputTopic));
-    configs.put("app.messageCount", String.valueOf(messageCount));
-    configs.put("app.outputTopic", outputTopic);
-    configs.put("app.outputSystem", testSystem);
-    configs.put(ZkConfig.ZK_CONNECT, zkConnect());
-
-    configs.put("job.systemstreampartition.grouper.factory",
-        "org.apache.samza.container.grouper.stream.GroupByPartitionFactory");
-    configs.put("task.name.grouper.factory", "org.apache.samza.container.grouper.task.GroupByContainerIdsFactory");
-
-    configs.put(JobCoordinatorConfig.JOB_COORDINATOR_FACTORY, "org.apache.samza.zk.ZkJobCoordinatorFactory");
-
-    return configs;
-  }
-
-  /**
-   * Produces the provided number of messages to the topic.
-   */
-  private void produceMessages(final int start, String topic, int numMessages) {
-    KafkaProducer producer = getKafkaProducer();
-    for (int i = start; i < numMessages + start; i++) {
-      try {
-        System.out.println("producing " + i);
-        producer.send(new ProducerRecord(topic, String.valueOf(i).getBytes())).get();
-      } catch (InterruptedException | ExecutionException e) {
-        e.printStackTrace();
-      }
-    }
-  }
-
-  /**
-   * Runs the provided stream processor by starting it, waiting on the provided latch with a timeout,
-   * and then stopping it.
-   */
-  private Thread runInThread(final StreamProcessor processor, CountDownLatch latch) {
-    Thread t = new Thread() {
-
-      @Override
-      public void run() {
-        processor.start();
-        try {
-          // just wait
-          synchronized (this) {
-            this.wait(100000);
-          }
-          System.out.println("notified. Abandon the wait.");
-        } catch (InterruptedException e) {
-          System.out.println("wait interrupted" + e);
-        }
-        System.out.println("Stopping the processor");
-        processor.stop();
-      }
-    };
-    return t;
-  }
-
-  // for sequential values we can generate them automatically
-  private void verifyNumMessages(String topic, int numberOfSequentialValues, int exectedNumMessages) {
-    // we should get each value one time
-    // create a map of all expected values to validate
-    Map<Integer, Boolean> expectedValues = new HashMap<>(numberOfSequentialValues);
-    for (int i = 0; i < numberOfSequentialValues; i++) {
-      expectedValues.put(i, false);
-    }
-    verifyNumMessages(topic, expectedValues, exectedNumMessages);
-  }
-
-  /**
-   * Consumes data from the topic until there are no new messages for a while
-   * and asserts that the number of consumed messages is as expected.
-   */
-  private void verifyNumMessages(String topic, final Map<Integer, Boolean> expectedValues, int expectedNumMessages) {
-    KafkaConsumer consumer = getKafkaConsumer();
-    consumer.subscribe(Collections.singletonList(topic));
-
-    Map<Integer, Boolean> map = new HashMap<>(expectedValues);
-    int count = 0;
-    int emptyPollCount = 0;
-
-    while (count < expectedNumMessages && emptyPollCount < 5) {
-      ConsumerRecords records = consumer.poll(5000);
-      if (!records.isEmpty()) {
-        Iterator<ConsumerRecord> iterator = records.iterator();
-        while (iterator.hasNext()) {
-          ConsumerRecord record = iterator.next();
-          String val = new String((byte[]) record.value());
-          System.out.println("Got value " + val);
-          map.put(Integer.valueOf(val), true);
-          count++;
-        }
-      } else {
-        emptyPollCount++;
-        System.out.println("empty polls " + emptyPollCount);
-      }
-    }
-    // filter out numbers we did not get
-    long numFalse = map.values().stream().filter(v -> !v).count();
-    Assert.assertEquals("didn't get this number of events ", 0, numFalse);
-    Assert.assertEquals(expectedNumMessages, count);
-  }
-
-  // StreamTaskClass
-  public static class TestStreamTask implements StreamTask, InitableTask {
-    // static field since there's no other way to share state b/w a task instance and
-    // stream processor when constructed from "task.class".
-    static CountDownLatch endLatch;
-    private int processedMessageCount = 0;
-    private String processorId;
-    private String outputTopic;
-    private String outputSystem;
-
-    @Override
-    public void init(Config config, TaskContext taskContext)
-        throws Exception {
-      this.processorId = config.get(ApplicationConfig.PROCESSOR_ID);
-      this.outputTopic = config.get("app.outputTopic", "output");
-      this.outputSystem = config.get("app.outputSystem", "test-system");
-    }
-
-    @Override
-    public void process(IncomingMessageEnvelope incomingMessageEnvelope, MessageCollector messageCollector,
-        TaskCoordinator taskCoordinator)
-        throws Exception {
-
-      Object message = incomingMessageEnvelope.getMessage();
-
-      messageCollector.send(new OutgoingMessageEnvelope(new SystemStream(outputSystem, outputTopic), message));
-      processedMessageCount++;
-
-      System.out.println(
-          "Stream processor " + processorId + "; offset=" + incomingMessageEnvelope.getOffset() + "; totalRcvd=" + processedMessageCount
-              + ";received " + message + "; ssp=" + incomingMessageEnvelope.getSystemStreamPartition());
-      synchronized (endLatch) {
-        endLatch.countDown();
-      }
-    }
   }
 }
